// src/client/models/TimeClockGroupViewModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TimeClockViewModel } from './TimeClockViewModel';
export type TimeClockGroupViewModel = {
    RecordDate?: string;
    RecordTime?: string;
    Clocks?: Array<TimeClockViewModel>;
};