// src/client/models/TimeEntryGroupViewModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TimeEntryDisplayModel } from './TimeEntryDisplayModel';
export type TimeEntryGroupViewModel = {
    RecordDate?: string;
    Entries?: Array<TimeEntryDisplayModel>;
};